仮設エディタにHTMLを文字列へ変換する処理を追加
id属性ごとに区切る機能をビルドスクリプトと共通化。
